# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BSY MIS Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerApplicant
      tags: [auth]
      summary: Register new applicant with mobile OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/send-otp:
    post:
      operationId: sendOtp
      tags: [auth]
      summary: Send OTP to mobile number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOtpInput"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpSentResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpInput"
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: staffLogin
      tags: [auth]
      summary: Staff login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /meetings:
    get:
      operationId: listMeetings
      tags: [meetings]
      summary: List PWC meetings
      responses:
        "200":
          description: Meetings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Meeting"
    post:
      operationId: scheduleMeeting
      tags: [meetings]
      summary: Schedule a PWC meeting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingInput"
      responses:
        "201":
          description: Meeting scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}:
    get:
      operationId: getMeeting
      tags: [meetings]
      summary: Get meeting by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Meeting
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /meetings/{id}/decision:
    post:
      operationId: recordMeetingDecision
      tags: [meetings]
      summary: Record PWC meeting decision
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DecisionInput"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meeting"
  /beneficiaries:
    get:
      operationId: listBeneficiaries
      tags: [beneficiaries]
      summary: List beneficiaries
      parameters:
        - name: sansthaId
          in: query
          schema:
            type: ["integer", "null"]
        - name: aadhaarSeeded
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Beneficiaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Beneficiary"
  /beneficiaries/{id}:
    get:
      operationId: getBeneficiary
      tags: [beneficiaries]
      summary: Get beneficiary by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Beneficiary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Beneficiary"
  /beneficiaries/{id}/payments:
    get:
      operationId: getBeneficiaryPayments
      tags: [payments]
      summary: Get payment history for beneficiary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentRecord"
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        mobile:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [applicant, facilitator, sanstha, gimaba, po, pwc]
        district:
          type: ["string", "null"]
        sansthaId:
          type: ["integer", "null"]
        status:
          type: string
        createdAt:
          type: string
      required: [id, mobile, name, role, status, createdAt]
    RegisterInput:
      type: object
      properties:
        mobile:
          type: string
        name:
          type: string
        otp:
          type: string
      required: [mobile, name, otp]
    SendOtpInput:
      type: object
      properties:
        mobile:
          type: string
      required: [mobile]
    OtpSentResponse:
      type: object
      properties:
        message:
          type: string
        expiresIn:
          type: integer
      required: [message, expiresIn]
    VerifyOtpInput:
      type: object
      properties:
        mobile:
          type: string
        otp:
          type: string
      required: [mobile, otp]
    StaffLoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
      required: [username, password, role]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        message:
          type: string
      required: [user, token, message]
...
    Meeting:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        scheduledDate:
          type: string
        mode:
          type: string
          enum: [online, offline]
        venueOrLink:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        decision:
          type: ["string", "null"]
          enum: [approved, rejected, null]
        decisionReason:
          type: ["string", "null"]
        minutes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, applicationId, scheduledDate, mode, venueOrLink, status, createdAt]
    MeetingInput:
      type: object
      properties:
        applicationId:
          type: integer
        scheduledDate:
          type: string
        mode:
          type: string
          enum: [online, offline]
        venueOrLink:
          type: string
      required: [applicationId, scheduledDate, mode, venueOrLink]
    DecisionInput:
      type: object
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        decisionReason:
          type: string
        minutes:
          type: ["string", "null"]
      required: [decision, decisionReason]
    Beneficiary:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        childName:
          type: string
        district:
          type: string
        approvalDate:
          type: string
        sansthaId:
          type: ["integer", "null"]
        sansthaName:
          type: ["string", "null"]
        beneficiaryCategory:
          type: string
        monthlyAmount:
          type: number
        status:
          type: string
          enum: [active, inactive, aged_out, suspended]
        aadhaarSeeded:
          type: boolean
        kycDone:
          type: boolean
        eligibilityRisk:
          type: boolean
        renewalDue:
          type: boolean
      required: [id, applicationId, childName, district, approvalDate, beneficiaryCategory, monthlyAmount, status, aadhaarSeeded, kycDone, eligibilityRisk, renewalDue]
    PaymentRecord:
      type: object
      properties:
        id:
          type: integer
        beneficiaryId:
          type: integer
        beneficiaryName:
          type: string
        paymentMonth:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [success, failed, pending, processing]
        transactionRef:
          type: ["string", "null"]
        failureReason:
          type: ["string", "null"]
        processedAt:
          type: ["string", "null"]
      required: [id, beneficiaryId, beneficiaryName, paymentMonth, amount, status]
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [info, success, warning, error]
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, userId, title, message, type, isRead, createdAt]
...