# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BSY MIS Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  mobile: string;
  name: string;
  role: UserRole;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  sansthaId?: number | null;
  status: string;
  createdAt: string;
}